requirements.txt
fastapi
uvicorn
celery[redis]
python-dotenv
flower